backend/requirements.txt
Django==4.2
djangorestframework==3.16.1
django-cors-headers==4.7.0
python-decouple==3.8
Pillow==10.0.0
